+++ server/pipeline_match.go
@@ -211,7 +211,7 @@ func (p *Pipeline) matchJoin(logger *zap.Logger, session Session, envelope *rtap
 	}
 
 	// Whether the user has just (successfully) joined the match or was already a member, return the match info anyway.
-	ps := p.tracker.ListByStream(stream, true)
+	ps := p.tracker.ListByStream(stream, false, true)
 	presences := make([]*rtapi.UserPresence, 0, len(ps))
 	for _, p := range ps {
 		if isNew && p.UserID == session.UserID() && p.ID.SessionID == session.ID() {